+++ BackEnd/Models/EnvironmentSpecificSettings.cs
@@ -36,5 +36,9 @@
         public string SqlDbConnectionString { get; set; }
 
         public string FrontEndUrl { get; set; }
+
+        public string ApplicationInsightsConnectionString { get; set; }
+
+        public string ApplicationInsightsInstrumentationKey { get; set; }
     }   
 }